requirements.txt
requests
beautifulsoup4
openpyx